requirements.txt
annotated-types
anyio
assemblyai
blinker
cachetools
certifi
charset-normalizer
click
colorama
exceptiongroup
flask
flask-login
flask-sqlalchemy
flask-wtf
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
importlib-metadata
itsdangerous
jinja2
markdown
markupsafe
numpy
pandas
pillow
proto-plus
protobuf
pyasn1
pyasn1-modules
pydantic
pydantic-core
pyparsing
python-dateutil
pytz
requests
rsa
six
sniffio
sqlalchemy
tqdm
typing-extensions
tzdata
uritemplate
urllib3
websockets
werkzeug
wtforms
zipp
